requirements.txt
Pillow>=3.3.1
piexif>=1.0.5
